feat: protected organistation routes

+++ server/src/controllers/organisation.ts
@@ -1,6 +1,8 @@
 import { Request, Response } from 'express';
 import Organisation from '../models/organisation';
 import Course from '../models/course';
+import { RequestWithUser } from '../types';
+import User from '../models/user';
 
 async function addOrganisation(req: Request, res: Response) {
   try {
@@ -11,8 +13,11 @@ async function addOrganisation(req: Request, res: Response) {
       return res.status(409).send({ message: 'Organisation name exists' });
     }
 
-    const ownerId = 'placeholder'; //placeholder before oauth
+    const ownerId = (req as RequestWithUser).user.id;
     const newOrg = await Organisation.createOrganisation(name, ownerId);
+
+    await User.addOrgToUser(ownerId, newOrg.id);
+    await User.assignRoleToUser(ownerId, newOrg.roles[0]);
 
     res.status(201).send(newOrg);
   } catch (error) {
@@ -46,7 +51,11 @@ async function editOrganisation(req: Request, res: Response) {
   try {
     const { orgId } = req.params;
     const newData = req.body;
-    const updatedOrg = await Organisation.editOrganisation(orgId, newData);
+
+    const userId = (req as RequestWithUser).user.id;
+    const updatedOrg = await Organisation.editOrganisation(orgId, newData, userId);
+    //TODO: Make a more descriptive error saying something like you can't delete this when the 
+    //userId is different to the orgs owner but for now this is enough
     res.status(200).send(updatedOrg);
   } catch (error) {
     console.log(error);
@@ -57,6 +66,13 @@ async function editOrganisation(req: Request, res: Response) {
 async function deleteOrganisation(req: Request, res: Response) {
   try {
     const { orgId } = req.params;
+
+    const ownerId = (req as RequestWithUser).user.id;
+    if(!await User.userIsOrgOwner(ownerId, orgId)) {
+      console.log('NOT ALLOWED')
+      return res.status(401).send({message: 'Unauthorised'})
+    }
+
     await Course.deleteCoursesInOrganisation(orgId);
     const deletedOrg = await Organisation.deleteOrganisation(orgId);
     res.status(204).send(deletedOrg);
